feat: Add the support of cloudwatch and prometheus
This commit provide a really opinionated configuration that we use at cycloid.
By default, it will expose a prometheus metrics endpoint, and it supports also a cloudwatch export.

It's a start of a simple log configuration, the goal is to use a maximum default value and to focus mainly on filenames and path.