fix: Change some parameters to avoid errors
Multiple things have been changed.
It's now running only one worker because cloudwatch and tail aren't able to use multiple workers.
There was a typo on default file list to watch, it needed an empty list.
There was also a problem with cloudwatch output plugin when it's running on multiple nodes. It needs to retry with a new token, it's not a problem as a PR already implemented this feature, but it was spamming the log file, so this commit add an error log level only on cloudwatch plugin.
This commit also fix the install and execution as root user.