systemd: Fix the systemd installation
When td-agent is setup on a systemd service manager, it fails to configure itself to run as root.
Because the service file is located in /lib and not in /etc. Systemd populate itself the /etc/systemd folder based on the /lib/systemd folder.

So during the install, the file in /etc doesn't exist yet.

This patch should fix the issue by modifying directly the service at the source.